refactor(client): extract SpaceSettingsState from SpacesComponent (A17.8b, part 1)
First half of A17.8. The settings-dialog state — openSettings, buildMeta, the
type-schema helpers, the duplicate-rule helpers and all four tabs' fields — moves
into a SpaceSettingsState service provided by the component. This is the
prerequisite for turning the tabs into child components: openSettings populates
all four tabs atomically and buildMeta reads across two of them, so the tabs
cannot each own their slice without one reaching into another. A service is the
seam. spaces.component.ts: 1893 -> 1616.

Move, don't rename. Member names are unchanged on purpose, so every moved method
body could be diffed against the original: 20 of 23 are byte-identical, and the
other 3 differ only by dropping `as TypeSchemaState & { _libRef?: string }` casts
made redundant by declaring _libRef on the interface (compile-time only).

That diff caught three defects my own hand-copy introduced, none of which the
characterization tests noticed:
  - addDupeRule defaulted to minScore 0.92, not 0.95
  - the dupeSaved.set(false) reset was dropped from addDupeRule AND removeDupeRule
  - wipeStatCols hardcoded English labels instead of
    transloco.translate('spaces.stats.*') — an i18n regression that builds clean
    and would have shipped silently
Renaming during the move would have hidden all three. Two new tests pin the
dupe-rule defaults the characterization suite had not asserted.

Tests moved with the code they cover, not rewritten to make this pass: the 36
assertions from #237 are unchanged, 22 now live in
space-settings-state.service.spec.ts (openSettings/buildMeta/$ref round-trip/dupe
rules) and 16 stay on the component (sortedSpaces/storageInfo/proxy/rendering).
16 + 22 = 38 = 36 relocated + 2 new. The library $ref round-trip — the one that
silently unlinks a schema if the _libRef sentinel is lost — is intact.

Part 2 will extract the create dialog and the four tabs as components.

Verified: client build clean; client suite 107 -> 109, all green.
Docs checked — no doc references the component internals and no UI behavior
changed; no update needed.
